@@ -2,11 +2,15 @@
 
 namespace Time2Split\Help\Classes;
 
+use \Time2Split\Help\Exception\UnmodifiableException;
+
 /**
  * Create a backed unmodifiable instance.
  *
  * @author Olivier Rodriguez (zuri)
  * @package time2help\class
+ * 
+ * @phpstan-template T
  */
 interface GetUnmodifiable
 {
@@ -18,10 +22,11 @@ interface GetUnmodifiable
      * 
      * Any operation modifying the content of the instance must throws an exception.
      * 
-     * @return self
-     *      A wrapper arround the object.
-     * @throws \Time2Split\Help\Exception\UnmodifiableException
-     *      If a mutable method is called.
+     * @phpstan-return T&IsUnmodifiable
+     * @return mixed
+     *      A {@see IsUnmodifiable} wrapper arround the object.
+     *      The instance must throw a {@see UnmodifiableException}
+     *      if a writing method is called.
      */
-    public function unmodifiable(): self;
+    public function unmodifiable(): mixed;
 }
