@@ -8,34 +8,37 @@ namespace Time2Split\Help\Container;
  * An implementation is provided by the trait ArrayAccessUtilities.
  * 
  * @author Olivier Rodriguez (zuri)
- * @package time2help\container
+ * @package time2help\container\interface
+ * 
+ * @template K
+ * @template V
  */
 interface ArrayAccessUpdating
 {
     /**
      * Updates some existant entries and add the unexistant ones.
      *
-     * @param iterable<T> ...$listsOfEntries
+     * @param iterable<K,V> ...$listsOfEntries
      *            Lists of entries to update.
-     * @return static This set.
+     * @return static<K,V> This set.
      */
     public function updateEntries(iterable ...$listsOfEntries): static;
 
     /**
      * Deletes some entries (by their keys).
      *
-     * @param iterable<T> ...$keys
+     * @param K ...$keys
      *            The keys to drop.
-     * @return static This set.
+     * @return static<K,V> This set.
      */
     public function unsetMore(mixed ...$keys): static;
 
     /**
      * Deletes some entries (by their keys).
      *
-     * @param iterable<T> ...$lists
+     * @param iterable<int,K> ...$listsOfKeys
      *            Lists of keys to drop.
-     * @return static This set.
+     * @return static<K,V> This set.
      */
     public function unsetFromList(iterable ...$listsOfKeys): static;
 }
