@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Time2Split\Help\Container;
 
-use Time2Split\Help\Container\Trait\ArrayAccessUpdating;
-use Time2Split\Help\Container\Trait\ArrayAccessWithStorage;
-use Time2Split\Help\Container\Trait\IteratorAggregateWithArrayStorage;
+use IteratorAggregate;
 use Time2Split\Help\Iterables;
 
 /**
@@ -35,30 +33,68 @@ use Time2Split\Help\Iterables;
  * 
  * @see https://www.php.net/manual/en/ref.array.php
  * @author Olivier Rodriguez (zuri)
- * @package time2help\container
+ * @package time2help\container\abstract
+ * 
+ * 
+ * @template K
+ * @template V
+ * @implements ContainerAA<K,V,ArrayContainer<K,V>,K,V>
+ * @implements ArrayAccessUpdating<K,V>
+ * @implements IteratorAggregate<K,V>
  */
 abstract class ArrayContainer
-extends ContainerWithArrayStorage
 implements
-    ArrayAccessContainer,
-    FetchingOpened
+    ContainerAA,
+    ArrayAccessUpdating,
+    ContainerPutMethods,
+    FetchingClosed,
+    IteratorAggregate
 {
+    /**
+     * @use Trait\ArrayAccessPutValue<V>
+     * @use Trait\ArrayAccessUpdating<K,V>
+     * @use Trait\ArrayAccessWithStorage<K,V>
+     * @use Trait\FetchingClosed<K,V,ArrayContainer<K,V>>
+     * @use Trait\IteratorAggregateWithStorage<K,V>
+     * @use Trait\IteratorToArray<K,V>
+     * @use Trait\ToArrayToArrayContainer<K,V>
+     */
     use
-        ArrayAccessUpdating,
-        ArrayAccessWithStorage,
-        Trait\FetchingOpened,
-        IteratorAggregateWithArrayStorage;
+        Trait\ArrayAccessPutValue,
+        Trait\ArrayAccessUpdating,
+        Trait\ArrayAccessWithStorage,
+        Trait\CountableWithStorage,
+        Trait\FetchingClosed,
+        Trait\IteratorAggregateWithStorage,
+        Trait\IteratorToArray,
+        Trait\ToArrayToArrayContainer;
+
+    public function __construct(
+        protected array $storage
+    ) {}
+
+    #[\Override]
+    public function copy(): static
+    {
+        return new static($this->storage);
+    }
 
     #[\Override]
     public function unmodifiable(): self
     {
-        return ArrayContainers::Unmodifiable($this);
+        return ArrayContainers::unmodifiable($this);
     }
 
     #[\Override]
     public static function null(): self
     {
         return ArrayContainers::null();
+    }
+
+    #[\Override]
+    public function clear(): void
+    {
+        $this->storage = [];
     }
 
     #[\Override]
@@ -75,7 +111,7 @@ implements
 
     #[\Override]
     public function equals(
-        ArrayContainer $other,
+        ContainerBase $other,
         bool|callable $strictOrEquals = false
     ): bool {
         if ($this === $other)
@@ -97,12 +133,12 @@ implements
 
     #[\Override]
     public function isIncludedIn(
-        ArrayContainer $other,
+        ContainerBase $other,
         bool|callable $strictOrEquals = false,
         bool $strictInclusion = false,
     ): bool {
         if ($strictInclusion)
-            return $this->isStrictlyIncludedIn($other, $strictOrEquals);
+            return $this->isStrictlyIncludedIn($other);
         if ($this === $other)
             return true;
 
@@ -129,6 +165,7 @@ implements
                         goto found;
                 }
                 return false;
+                /* @phpstan-ignore deadCode.unreachable */
                 found:
                 unset($b[$kb]);
                 return true;
@@ -155,6 +192,7 @@ implements
         if (!isset($theFunction))
             goto error;
 
+        /* @phpstan-ignore variable.undefined */
         $reflect = new \ReflectionFunction($theFunction);
         $return = $reflect->getReturnType();
 
@@ -171,9 +209,12 @@ implements
             $preArguments[] = array_shift($arguments);
         }
 
+        /* @phpstan-ignore variable.undefined */
         if ($isRef)
+            /* @phpstan-ignore variable.undefined */
             $ret = $theFunction(...[...$preArguments, &$this->storage, ...$arguments]);
         else
+            /* @phpstan-ignore variable.undefined */
             $ret = $theFunction(...[...$preArguments, $this->storage, ...$arguments]);
 
         if ((string)$return !== 'array')
@@ -181,6 +222,7 @@ implements
 
         $this->storage = $ret;
         return $this;
+        /* @phpstan-ignore deadCode.unreachable */
         error:
         throw new \BadFunctionCallException('Function ' . __CLASS__ . '::' . $name . ' is not callable');
     }
