@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Time2Split\Help\Container;
 
+use Closure;
 use Time2Split\Help\Classes\IsUnmodifiable;
 use Time2Split\Help\Classes\NotInstanciable;
 use Time2Split\Help\Container\ArrayContainer;
 use Time2Split\Help\Container\Trait\ContainerMapKey;
 use Time2Split\Help\Container\Trait\IteratorToArrayOfEntries;
-use Time2Split\Help\Container\Trait\UnmodifiableArrayAccessContainer;
 use Time2Split\Help\Iterables;
 
 /**
@@ -22,10 +22,12 @@ final class ArrayContainers
     use NotInstanciable;
 
     /**
+     * @template K
+     * @template V
      * 
-     * @param iterable ...$arrays
+     * @param iterable<K,V> ...$arrays
      *       The initial array contents
-     * @return ArrayContainer A new array container.
+     * @return ArrayContainer<K,V> A new array container.
      */
     static public function create(iterable ...$arrays)
     {
@@ -41,15 +43,19 @@ final class ArrayContainers
      * This class permits to handle more types of values and not just array keys.
      * It makes a bijection between a valid array key and an element.
      *
-     * @param callable $mapKey
+     * @template K
+     * @template KMAP
+     * @template V
+     * 
+     * @param Closure(K):KMAP $mapKey
      *       Map an input item to a valid key.
-     * @param iterable ...$arrays
+     * @param iterable<K,V> ...$iterables
      *       The initial array contents
-     * @return ArrayContainer A new array container.
+     * @return ArrayContainer<K,V> A new array container.
      */
-    static public function toArrayKeys(callable $mapKey, iterable ...$arrays)
+    static public function toArrayKeys(Closure $mapKey, iterable ...$iterables): ArrayContainer
     {
-        $array = \iterator_to_array(Iterables::append(...$arrays));
+        $array = \iterator_to_array(Iterables::append(...$iterables));
 
         return new class($mapKey, $array) extends ArrayContainer {
             use ContainerMapKey,
@@ -71,35 +77,35 @@ final class ArrayContainers
                 $ret->copyMapKeyInternals($this);
                 return $ret;
             }
-            #[\Override]
-            public function toArrayContainer(): ArrayContainer
-            {
-                return $this->copy();
-            }
         };
     }
 
     static public function null(): ArrayContainer
     {
-        static $null = new class([])
-        extends ArrayContainer
-        implements IsUnmodifiable
-        {
-            use UnmodifiableArrayAccessContainer;
-        };
+        static $null = self::unmodifiable(ArrayContainers::create());
         return $null;
     }
 
-    static public function unmodifiable(ArrayContainer $subject): ArrayContainer
+    static public function unmodifiable(ArrayContainer $subject): ArrayContainer&IsUnmodifiable
     {
         return new class($subject)
         extends ArrayContainer
         implements IsUnmodifiable
         {
-            use UnmodifiableArrayAccessContainer;
+            use
+                Trait\UnmodifiableArrayAccess,
+                Trait\UnmodifiableArrayAccessUpdating,
+                Trait\UnmodifiableContainerPutMethods,
+                Trait\UnmodifiableClearable;
             public function __construct(ArrayContainer $subject)
             {
                 $this->storage = &$subject->storage;
+            }
+
+            #[\Override]
+            public function copy(): static
+            {
+                return $this;
             }
         };
     }
