@@ -8,6 +8,6 @@ namespace Time2Split\Help\Container;
  * A class implementing this interface must use the trait of the same name ({@see Trait\FetchingClosed}).
  * 
  * @author Olivier Rodriguez (zuri)
- * @package time2help\container
+ * @package time2help\container\interface
  */
 interface FetchingClosed {}
