@@ -4,18 +4,18 @@ declare(strict_types=1);
 
 namespace Time2Split\Help\Container;
 
+use Time2Split\Help\Cast\Cast;
 use Time2Split\Help\Classes\IsUnmodifiable;
 use Time2Split\Help\Classes\NotInstanciable;
 use Time2Split\Help\Container\_internal\SetWithStorage;
-use Time2Split\Help\Container\Container;
 use Time2Split\Help\Container\Trait\IteratorToArray;
 use Time2Split\Help\Container\Trait\IteratorToArrayContainer;
-use Time2Split\Help\Container\Trait\UnmodifiableArrayAccessContainer;
+use Time2Split\Help\Container\Trait\UnmodifiableContainerAA;
 use Time2Split\Help\Container\Trait\UnmodifiableContainerPutMethods;
 use Time2Split\Help\Iterables;
 
 /**
- * Factories and functions on set.
+ * Factories and functions on sets.
  * 
  * @package time2help\container
  * @author Olivier Rodriguez (zuri)
@@ -24,14 +24,16 @@ final class Sets
 {
     use NotInstanciable;
 
-    private static function create(Container $storage): Set
+    private static function create(ContainerAA $storage): Set
     {
         return new class($storage)
         extends SetWithStorage {
             #[\Override]
             public function getIterator(): \Traversable
             {
-                return Iterables::keys(parent::getIterator());
+                return Cast::iterableToIterator(
+                    Cast::iterableToIterator(Iterables::keys(parent::getIterator()))
+                );
             }
         };
     }
@@ -95,7 +97,7 @@ final class Sets
                 IteratorToArray,
                 IteratorToArrayContainer;
             public function __construct(
-                Container $storage,
+                ContainerAA $storage,
                 private readonly string $enumClass
             ) {
                 parent::__construct($storage);
@@ -120,7 +122,7 @@ final class Sets
             #[\Override]
             public function getIterator(): \Traversable
             {
-                return Iterables::keys(parent::getIterator());
+                return Cast::iterableToIterator(Iterables::keys(parent::getIterator()));
             }
         };
     }
@@ -153,15 +155,16 @@ final class Sets
      * 
      * @param Set<T> $set
      *            A set to decorate.
-     * @return Set<T> The backed unmodifiable set.
+     * @return SetWithStorage<T> The backed unmodifiable set.
      */
     public static function unmodifiable(Set $set): Set
     {
-        return new class($set)
+        assert($set instanceof SetWithStorage);
+        return new class($set->getStorage())
         extends SetWithStorage
         implements IsUnmodifiable
         {
-            use UnmodifiableArrayAccessContainer,
+            use UnmodifiableContainerAA,
                 UnmodifiableContainerPutMethods;
         };
     }
@@ -171,9 +174,9 @@ final class Sets
      *
      * The value is a singleton and may be compared with the `===` operator.
      * 
-     * @return Set<void> The unique null pattern set.
+     * @return SetWithStorage<void> The unique null pattern set.
      */
-    public static function null(): Set
+    public static function null(): SetWithStorage
     {
         static $null = self::unmodifiable(self::arrayKeys());
         return $null;
