@@ -10,27 +10,19 @@ use Time2Split\Help\Container\ArrayContainers;
 /**
  * An implementation of a Container using an internal array storage.
  * 
- * ```
- * public function offsetExists(mixed $offset): bool
- * {
- *     return \array_key_exists($offset, $this->storage);
- * }
- * public function clear(): void
- * {
- *     $this->storage = [];
- * }
- * public function toArray(): array
- * {
- *     return $this->storage;
- * }
- * ```
- * @var array $storage The internal storage must be defined into the class.
- * 
  * @author Olivier Rodriguez (zuri)
  * @package time2help\container
+ * 
+ * @template K
+ * @template V
+ * 
+ * @property array $storage The internal storage must be defined into the class definition.
  */
 trait ContainerWithArrayStorage
 {
+    /**
+     * @use IteratorAggregateWithArrayStorage<K,V>
+     */
     use
         CountableWithStorage,
         IteratorAggregateWithArrayStorage;
@@ -41,15 +33,22 @@ trait ContainerWithArrayStorage
         $this->storage = [];
     }
 
+    /**
+     * @return array<K,V>
+     */
     #[\Override]
     public function toArray(): array
     {
         return $this->storage;
     }
 
+    /**
+     * @return ArrayContainer<K,V>
+     */
     #[\Override]
     public function toArrayContainer(): ArrayContainer
     {
+        /* @phpstan-ignore return.type */
         return ArrayContainers::create($this->storage);
     }
 }
