@@ -2,33 +2,75 @@
 
 namespace Time2Split\Help\Container\Trait;
 
+use Closure;
+use Time2Split\Help\Container\ContainerBase;
+
 /**
  * Methods for simple querying.
  *
- * It must be used in the first effective object class declaration to set
- * properly the `self` typyng contraints.
- * 
  * @author Olivier Rodriguez (zuri)
  * @package time2help\container
+ * 
+ * @template K
+ * @template V
+ * @template C of ContainerBase<K,V>
  */
 trait FetchingOpened
 {
+    /**
+     * Whether another container have the same contents according to
+     * a comparison strategy.
+     * 
+     * @param C $other The other container.
+     * @param bool|Closure $strictOrEquals The comparison strategy.
+     *  - `true`: The `===` operator is used.
+     *  - `false`: The `==` operator is used.
+     *  - `strictOrEquals(mixed $a, mixed $b):bool`
+     * 
+     *      Returns true if the items $a and $b are equals.
+     * 
+     * @return bool
+     * - `true`: the other container contains exactly the same elements.
+     * - `false`: otherwise.
+     */
     abstract public function equals(
         self $other,
-        bool|callable $strictOrEquals = false
+        bool|Closure $strictOrEquals = false
     ): bool;
 
-    abstract private function isIncludedIn(
+    /**
+     * Whether another container included the contents of $this according to
+     * a comparison strategy.
+     * 
+     * @param C $other The other container.
+     * @param bool|Closure $strictOrEquals The comparison strategy.
+     *  - `true`: The `===` operator is used.
+     *  - `false`: The `==` operator is used.
+     *  - `strictOrEquals(mixed $a, mixed $b):bool`
+     * 
+     *      Returns true if the items $a and $b are equals.
+     * @param $strictInclusion
+     *  - `true`: The inclusion must be strict (the other container have more elements)
+     *  - `false`: The inclusion is not necessary strict (the other container may be equal)
+     * 
+     * @return bool
+     * - `true`: the other container contains the same elements (and maybe more).
+     * - `false`: otherwise.
+     */
+    abstract public function isIncludedIn(
         self $other,
-        bool|callable $strictOrEquals = false,
+        bool|Closure $strictOrEquals = false,
         bool $strictInclusion = false,
     ): bool;
 
     // ========================================================================
 
+    /**
+     * @param C $other
+     */
     protected final function isStrictlyIncludedIn(
         self $other,
-        bool|callable $strictOrEquals = false,
+        bool|Closure $strictOrEquals = false,
     ): bool {
 
         if ($this === $other)
