@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace Time2Split\Help\Container\Trait;
 
+use ArrayIterator;
+use Iterator;
+use Traversable;
+
 /**
  * An implementation of \IteratorAggregate returning the internal array storage.
  * 
@@ -18,12 +22,18 @@ namespace Time2Split\Help\Container\Trait;
  * 
  * @author Olivier Rodriguez (zuri)
  * @package time2help\container
+ * 
+ * @template K
+ * @template V
  */
 trait IteratorAggregateWithArrayStorage
 {
+    /**
+     * @return Iterator<K,V>
+     */
     #[\Override]
-    public function getIterator(): \Traversable
+    public function getIterator(): Traversable
     {
-        return new \ArrayIterator($this->storage);
+        return new ArrayIterator($this->storage);
     }
 }
