@@ -3,7 +3,8 @@
 namespace Time2Split\Help\Container\Trait;
 
 /**
- * An implementation of `ToArray::toArray`.
+ * An implementation of `ToArray::toArray`
+ * transforming the instance's entries into an array.
  * 
  * ```
  * function toArray(): array
@@ -16,9 +17,15 @@ namespace Time2Split\Help\Container\Trait;
  * 
  * @author Olivier Rodriguez (zuri)
  * @package time2help\class
+ * 
+ * @template K
+ * @template V
  */
 trait IteratorToArray
 {
+    /**
+     * @return array<K,V>
+     */
     #[\Override]
     public function toArray(): array
     {
