@@ -17,12 +17,18 @@ use Time2Split\Help\Container\ArrayContainers;
  *
  * @author Olivier Rodriguez (zuri)
  * @package time2help\class
+ * 
+ * @template K
+ * @template V
  */
 trait IteratorToArrayContainer
 {
+    /**
+     * @return ArrayContainer<K,V>
+     */
     #[\Override]
     public function toArrayContainer(): ArrayContainer
     {
-        return ArrayContainers::create($this->getIterator());
+        return ArrayContainers::create($this);
     }
 }
