@@ -4,7 +4,6 @@ namespace Time2Split\Help\Container\Trait;
 
 use Time2Split\Help\Container\ArrayContainer;
 use Time2Split\Help\Container\ArrayContainers;
-use Time2Split\Help\Container\Entry;
 
 /**
  * An implementation of `ToArray::toArrayContainer`.
@@ -18,12 +17,21 @@ use Time2Split\Help\Container\Entry;
  *
  * @author Olivier Rodriguez (zuri)
  * @package time2help\class
+ * 
+ * @template K
+ * @template V
  */
 trait ToArrayToArrayContainer
 {
+    /**
+     * @return ArrayContainer<K,V>
+     */
     #[\Override]
     public function toArrayContainer(): ArrayContainer
     {
+        /**
+         * @var ArrayContainer<K,V>
+         */
         return ArrayContainers::create($this->toArray());
     }
 }
