@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Time2Split\Help\Tests\Container;
 
+use Closure;
 use PHPUnit\Framework\Assert;
 use Time2Split\Help\Arrays;
 use Time2Split\Help\Container\FetchingClosed;
-use Time2Split\Help\Container\Container;
 use Time2Split\Help\Container\ContainerPutMethods;
 use Time2Split\Help\Container\Entry;
 use Time2Split\Help\Container\FetchingOpened;
@@ -21,9 +21,9 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
 {
     protected const MIN_NB_ENTRIES = 6;
 
-    abstract protected static function provideContainer(): Container;
+    abstract protected static function provideContainer(): mixed;
     abstract protected static function provideEntries(): array;
-    abstract protected static function provideContainerWithSubEntries(int $offset = 0, ?int $length = null): Container;
+    abstract protected static function provideContainerWithSubEntries(int $offset = 0, ?int $length = null);
 
     // ========================================================================
 
@@ -46,6 +46,16 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
         return \array_slice(static::provideEntryObjects(), $offset, $length, true);
     }
 
+    protected static function entriesEqualClosure_traversableTest(bool $strict = false): Closure
+    {
+        return Entry::equalsClosure($strict);
+    }
+
+    protected static function entriesEqualClosure_putMethodTest(bool $strict = false): Closure
+    {
+        return Entry::equalsClosure($strict);
+    }
+
     // ========================================================================
 
     public static function setUpBeforeClass(): void
@@ -59,7 +69,7 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
     // ========================================================================
 
     #[\Override]
-    protected static final function provideSubject(): Container
+    protected static final function provideSubject(): mixed
     {
         return static::provideContainer();
     }
@@ -85,7 +95,7 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
         $subject = static::provideContainerWithSubEntries();
         $copy = $subject->copy();
         $this->checkNotEmpty($copy);
-        $this->checkListEquals($subject, $copy);
+        $this->checkEntriesAreEqual($subject, $copy);
     }
 
     final public function testClearableContainer(): void
@@ -108,13 +118,13 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
         $b = fn() => Entry::traverseListOfEntries(static::provideSubEntries(3, 3));
         $array = fn() => Iterables::append($a(), $b());
 
-        $subject->putMore(...Iterables::keys($a()));
+        $subject->putMore(...\iterator_to_array(Iterables::keys($a())));
         $this->checkNotEmpty($subject, 3);
-        $this->checkValuesEqualsProvidedEntries($subject, $a());
+        $this->checkEntriesAreEqual($subject, $a(), static::entriesEqualClosure_putMethodTest());
 
-        $subject->putMore(...Iterables::keys($b()));
+        $subject->putFromList(Iterables::keys($b()));
         $this->checkNotEmpty($subject, 6);
-        $this->checkEqualsProvidedEntries($subject, $array());
+        $this->checkEntriesAreEqual($subject, $array(), static::entriesEqualClosure_putMethodTest());
     }
 
     final public function testToArrayContainer(): void
@@ -129,7 +139,7 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
     {
         $subject = static::provideContainerWithSubEntries();
         $entries = fn() => Entry::traverseEntries($this->provideSubEntries());
-        $this->checkValuesEqualsProvidedEntries($subject, $entries());
+        $this->checkEntriesAreEqual($subject, $entries(), static::entriesEqualClosure_traversableTest());
     }
 
     // ========================================================================
@@ -328,6 +338,7 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
 
     final public function testFetchingClosedIncludesInContainerStrict(): void
     {
+        $this->markTestSkipped("disabled");
         $strict = true;
         /**
          * @var Trait\FetchingClosed
@@ -336,7 +347,6 @@ abstract class AbstractContainerTestClass extends AbstractClassesTestClass
         if (!($a instanceof FetchingClosed))
             $this->markTestSkipped();
 
-        $eq = fn($a, $b) => $a == $b;
         $this->assertFalse($a->isIncludedIn($a,  $strict), '(!) a <= a');
 
         /**
